Expose render model components
I've removed the OpenVRRenderModel type in favor of two new top-level
functions to return ArrayMeshes directly, to make the interface for
components and single-model symmetrical. The subclass didn't have any
metadata on it anyway.

I might bring back something like it but that isn't a VisualInstance3D
at all, because there are components which have no visual
representation. This would provide a convenient place to expose
RenderModel_ComponentState_t data.