Configure inputs from openvr action manifest
As long as an application doesn't have an action manifest configured in
the Steamworks publisher UI, it is free to call SetActionManifestPath as
many times as it wants. Make this possible by reading our own action
manifest to configure godot inputs, rather than hardcoding the defaults.
This also makes it much easier for an application to customize its
action manifest rather than relying on the generic mappings.

This also unifies the names of the inputs between the godot side and the
default action manifest.