Avoid removing the wrong trackers when detaching
When detaching a tracker which has become inactive, verify that it still
owns its name in the XRServer. If a newer tracker has replaced it, for
example when switching between controllers and hand tracking, attempting
to remove it will accidentally remove the newer one instead.

This should be fixed upstream, but for now this workaround avoids the
problem and will be forward compatible.